Feat: ShoppingList support (resolve #1 resolve #2)
Add basic UI elements such as an main activity with a sidebar and
fragments for showing data.

In addition to shopping-list specific fragments such as ManageLists
and ViewShoppingList, some general fragments to get basic functionality
is included, more specificly Settings, About and an onboarding fragment
to get users to create their first Shopping List. For now, the about
screen only contains the app name, and the settings screen can only
change the server address and port.